[rtl] Dedicated ibex_counter_flop modules
Depending on whether we are targeting a Xilinx FPGA or a generic
target, we use a Xilinx specifc flop implementation for the
Ibex counters. To increase readability, this commit creates a
new module containing a Xilinx specific flop implementation. For
generic targets, the existing prim_flop is used. In the Xilinx
specific version, a DSP is used when the CounterWidth is below
49-bits.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>